Fix waterboost persistence bug
At the moment, debugging from a map with waterboosting enabled to another map will keep waterboosting enabled. This is because LevelEnter is not called when debugging, so the grab meta is never checked again.

This fix just changes the check from LevelEnter to LevelLoader, which will be called for all level loading methods